nrzled.NewSPI: also add padding in the front
Some SPI devices take a while to set up for transfer, and while doing
so, show the first bit they're about to send.

This causes the first LED to always interpret the first symbol as a
logical 1. Zapping out the whole strip will leave the first LED green,
and the first pixel will always be slightly green-ish.

Prevent this, by applying the same 3 bit padding we have at the end to
the front.

Signed-off-by: Florian Klink <flokli@flokli.de>